+++ src/index.ts
@@ -1,6 +1,7 @@
-export * from "./safeHandler.js";
+export { safeHandler } from "./safeHandler.js";
 export * from "./camelize.js";
 export * from "./errors.js";
+export * as typeHelper from "./typeHelper.js";
 export * as validator from "./validator.js";
 export * as fs from "./fs.js";
 export * from "./types.js";